Move the location-related models into a separate namespace
