+++ common/internal_model/src/main/scala/uk/ac/wellcome/models/work/internal/Holdings.scala
@@ -1,5 +1,7 @@
 package uk.ac.wellcome.models.work.internal
 
+import weco.catalogue.internal_model.locations.PhysicalLocation
+
 case class Holdings(
   note: Option[String],
   enumeration: List[String],
