+++ common/internal_model/src/main/scala/weco/catalogue/internal_model/locations/LocationType.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.models.work.internal
+package weco.catalogue.internal_model.locations
 
 import enumeratum.{Enum, EnumEntry}
 import io.circe.{Decoder, Encoder}
