+++ common/internal_model/src/test/scala/uk/ac/wellcome/models/work/generators/ItemsGenerators.scala
@@ -1,8 +1,17 @@
 package uk.ac.wellcome.models.work.generators
 
 import uk.ac.wellcome.models.work.internal._
-import weco.catalogue.internal_model.generators.IdentifiersGenerators
+import weco.catalogue.internal_model.generators.{
+  IdentifiersGenerators,
+  LocationGenerators
+}
 import weco.catalogue.internal_model.identifiers.{IdState, SourceIdentifier}
+import weco.catalogue.internal_model.locations.{
+  AccessStatus,
+  License,
+  Location,
+  LocationType
+}
 
 trait ItemsGenerators extends IdentifiersGenerators with LocationGenerators {
