+++ common/internal_model/src/test/scala/weco/catalogue/internal_model/locations/LicenseTest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.models.work.internal
+package weco.catalogue.internal_model.locations
 
 import org.scalatest.Assertion
 import org.scalatest.funspec.AnyFunSpec
