+++ pipeline/inferrer/inference_manager/src/main/scala/uk/ac/wellcome/platform/inference_manager/services/ImageDownloader.scala
@@ -17,8 +17,8 @@ import akka.stream.{IOResult, Materializer}
 import akka.stream.scaladsl.{FileIO, Flow, Keep, Sink}
 import akka.util.ByteString
 import software.amazon.awssdk.services.sqs.model.Message
-import uk.ac.wellcome.models.work.internal.{DigitalLocation, LocationType}
 import uk.ac.wellcome.platform.inference_manager.models.DownloadedImage
+import weco.catalogue.internal_model.locations.{DigitalLocation, LocationType}
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.{Failure, Success, Try}
