+++ pipeline/inferrer/inference_manager/src/test/scala/uk/ac/wellcome/platform/inference_manager/services/ImageDownloaderTest.scala
@@ -10,7 +10,6 @@ import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
 import uk.ac.wellcome.akka.fixtures.Akka
 import uk.ac.wellcome.fixtures.TestWith
-import uk.ac.wellcome.models.work.internal.LocationType
 import uk.ac.wellcome.platform.inference_manager.fixtures.{
   MemoryFileWriter,
   RequestPoolFixtures,
@@ -19,6 +18,7 @@ import uk.ac.wellcome.platform.inference_manager.fixtures.{
 }
 import uk.ac.wellcome.platform.inference_manager.models.DownloadedImage
 import weco.catalogue.internal_model.generators.ImageGenerators
+import weco.catalogue.internal_model.locations.LocationType
 
 class ImageDownloaderTest
     extends AnyFunSpec
