+++ pipeline/transformer/transformer_calm/src/main/scala/uk/ac/wellcome/platform/transformer/calm/transformers/CalmItems.scala
@@ -1,15 +1,15 @@
 package uk.ac.wellcome.platform.transformer.calm.transformers
 
-import uk.ac.wellcome.models.work.internal.{
-  AccessCondition,
-  AccessStatus,
-  Item,
-  LocationType,
-  PhysicalLocation
-}
+import uk.ac.wellcome.models.work.internal.Item
 import uk.ac.wellcome.models.work.internal.result._
 import uk.ac.wellcome.platform.transformer.calm.CalmRecordOps
 import weco.catalogue.internal_model.identifiers.IdState
+import weco.catalogue.internal_model.locations.{
+  AccessCondition,
+  AccessStatus,
+  LocationType,
+  PhysicalLocation
+}
 import weco.catalogue.source_model.calm.CalmRecord
 
 object CalmItems extends CalmRecordOps {
