+++ pipeline/transformer/transformer_miro/src/main/scala/uk/ac/wellcome/platform/transformer/miro/transformers/MiroLocation.scala
@@ -1,12 +1,12 @@
 package uk.ac.wellcome.platform.transformer.miro.transformers
 
-import uk.ac.wellcome.models.work.internal.{
+import uk.ac.wellcome.platform.transformer.miro.source.MiroRecord
+import weco.catalogue.internal_model.locations.{
   AccessCondition,
   AccessStatus,
   DigitalLocation,
   LocationType
 }
-import uk.ac.wellcome.platform.transformer.miro.source.MiroRecord
 
 trait MiroLocation extends MiroLicenses with MiroContributorCodes {
