+++ pipeline/transformer/transformer_miro/src/main/scala/uk/ac/wellcome/platform/transformer/miro/transformers/MiroThumbnail.scala
@@ -1,7 +1,7 @@
 package uk.ac.wellcome.platform.transformer.miro.transformers
 
-import uk.ac.wellcome.models.work.internal.{DigitalLocation, LocationType}
 import uk.ac.wellcome.platform.transformer.miro.source.MiroRecord
+import weco.catalogue.internal_model.locations.{DigitalLocation, LocationType}
 
 trait MiroThumbnail extends MiroImageData with MiroLicenses {
   def getThumbnail(miroRecord: MiroRecord): DigitalLocation =
