+++ pipeline/transformer/transformer_miro/src/test/scala/uk/ac/wellcome/platform/transformer/miro/transformers/MiroLocationTest.scala
@@ -2,14 +2,13 @@ package uk.ac.wellcome.platform.transformer.miro.transformers
 
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
-import uk.ac.wellcome.models.work.internal.{
+import uk.ac.wellcome.platform.transformer.miro.generators.MiroRecordGenerators
+import weco.catalogue.internal_model.locations.{
   AccessCondition,
   AccessStatus,
   DigitalLocation,
-  License,
   LocationType
 }
-import uk.ac.wellcome.platform.transformer.miro.generators.MiroRecordGenerators
 
 class MiroLocationTest
     extends AnyFunSpec
