+++ pipeline/transformer/transformer_miro/src/test/scala/uk/ac/wellcome/platform/transformer/miro/transformers/MiroRecordTransformerCopyrightTest.scala
@@ -3,9 +3,9 @@ package uk.ac.wellcome.platform.transformer.miro.transformers
 import org.scalatest.Assertion
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
-import uk.ac.wellcome.models.work.internal.DigitalLocation
 import uk.ac.wellcome.platform.transformer.miro.generators.MiroRecordGenerators
 import uk.ac.wellcome.platform.transformer.miro.source.MiroRecord
+import weco.catalogue.internal_model.locations.DigitalLocation
 
 class MiroRecordTransformerCopyrightTest
     extends AnyFunSpec
