+++ pipeline/transformer/transformer_sierra/src/main/scala/uk/ac/wellcome/platform/transformer/sierra/transformers/SierraLocation.scala
@@ -8,6 +8,13 @@ import uk.ac.wellcome.platform.transformer.sierra.source.{
   SierraQueryOps,
   VarField
 }
+import weco.catalogue.internal_model.locations.{
+  AccessCondition,
+  AccessStatus,
+  LocationType,
+  PhysicalLocation,
+  PhysicalLocationType
+}
 import weco.catalogue.sierra_adapter.models.{SierraBibNumber, SierraItemNumber}
 
 trait SierraLocation extends SierraQueryOps with Logging {
