+++ pipeline/transformer/transformer_sierra/src/test/scala/uk/ac/wellcome/platform/transformer/sierra/transformers/SierraElectronicResourcesTest.scala
@@ -2,17 +2,17 @@ package uk.ac.wellcome.platform.transformer.sierra.transformers
 
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
-import uk.ac.wellcome.models.work.internal.AccessStatus.LicensedResources
-import uk.ac.wellcome.models.work.internal.{
-  AccessCondition,
-  DigitalLocation,
-  Item
-}
-import uk.ac.wellcome.models.work.internal.LocationType.OnlineResource
+import weco.catalogue.internal_model.locations.AccessStatus.LicensedResources
+import uk.ac.wellcome.models.work.internal.Item
+import weco.catalogue.internal_model.locations.LocationType.OnlineResource
 import uk.ac.wellcome.platform.transformer.sierra.generators.MarcGenerators
 import uk.ac.wellcome.platform.transformer.sierra.source.{
   MarcSubfield,
   VarField
+}
+import weco.catalogue.internal_model.locations.{
+  AccessCondition,
+  DigitalLocation
 }
 import weco.catalogue.sierra_adapter.generators.SierraGenerators
